# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZhuuVIP API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export interface LinkItem {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
  isActive: boolean;
}

export interface LinkInput {
  title: string;
  url: string;
  icon?: string;
  imageUrl?: string;
  sortOrder?: number;
}

export interface LinkUpdate {
  title?: string;
  url?: string;
  icon?: string;
  imageUrl?: string;
  sortOrder?: number;
  isActive?: boolean;
}

export interface Song {
  id: number;
  title: string;
  artist: string;
  url: string;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  duration?: number | null;
  sortOrder: number;
}

export interface SongInput {
  title: string;
  artist: string;
  url: string;
  coverUrl?: string;
  duration?: number;
  sortOrder?: number;
}

export interface SongUpdate {
  title?: string;
  artist?: string;
  url?: string;
  coverUrl?: string;
  duration?: number;
  sortOrder?: number;
}

export interface SiteSettings {
  profileName?: string;
  profileBio?: string;
  logoUrl?: string;
  bannerUrl?: string;
  themeColor?: string;
  statusText?: string;
}