# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ZhuuVIP API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: anthropic
    description: Anthropic AI chat operations
  - name: links
    description: Linktree link management
  - name: songs
    description: Music playlist management
  - name: settings
    description: Site settings management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /anthropic/conversations:
    get:
      operationId: listAnthropicConversations
      tags: [anthropic]
      summary: List all conversations for current user
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicConversation"
    post:
      operationId: createAnthropicConversation
      tags: [anthropic]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnthropicConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversation"

  /anthropic/conversations/{id}:
    get:
      operationId: getAnthropicConversation
      tags: [anthropic]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    delete:
      operationId: deleteAnthropicConversation
      tags: [anthropic]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"

  /anthropic/conversations/{id}/messages:
    get:
      operationId: listAnthropicMessages
      tags: [anthropic]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicMessage"
    post:
      operationId: sendAnthropicMessage
      tags: [anthropic]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnthropicMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}

  /links:
    get:
      operationId: listLinks
      tags: [links]
      summary: List all Linktree links
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinkItem"
    post:
      operationId: createLink
      tags: [links]
      summary: Create a new link (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkInput"
      responses:
        "201":
          description: Created link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkItem"

  /links/{id}:
    patch:
      operationId: updateLink
      tags: [links]
      summary: Update a link (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkUpdate"
      responses:
        "200":
          description: Updated link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkItem"
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete a link (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /songs:
    get:
      operationId: listSongs
      tags: [songs]
      summary: List all playlist songs
      responses:
        "200":
          description: List of songs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Song"
    post:
      operationId: createSong
      tags: [songs]
      summary: Add a new song (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongInput"
      responses:
        "201":
          description: Created song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"

  /songs/{id}:
    patch:
      operationId: updateSong
      tags: [songs]
      summary: Update a song (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongUpdate"
      responses:
        "200":
          description: Updated song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
    delete:
      operationId: deleteSong
      tags: [songs]
      summary: Delete a song (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all site settings (public)
      responses:
        "200":
          description: Settings as key-value object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AnthropicConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    AnthropicConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    AnthropicMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    AnthropicConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    AnthropicError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    LinkItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
        icon:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
      required:
        - id
        - title
        - url
        - sortOrder
        - isActive

    LinkInput:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        icon:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
      required:
        - title
        - url

    LinkUpdate:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        icon:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean

    Song:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        url:
          type: string
        coverUrl:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        sortOrder:
          type: integer
      required:
        - id
        - title
        - artist
        - url
        - sortOrder

    SongInput:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        url:
          type: string
        coverUrl:
          type: string
        duration:
          type: integer
        sortOrder:
          type: integer
      required:
        - title
        - artist
        - url

    SongUpdate:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        url:
          type: string
        coverUrl:
          type: string
        duration:
          type: integer
        sortOrder:
          type: integer

    SiteSettings:
      type: object
      properties:
        profileName:
          type: string
        profileBio:
          type: string
        logoUrl:
          type: string
        bannerUrl:
          type: string
        themeColor:
          type: string
        statusText:
          type: string