# lib/api-zod/src/generated/types/linkItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZhuuVIP API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LinkItem {
  id: number;
  title: string;
  url: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
  isActive: boolean;
}