# lib/api-zod/src/generated/types/linkUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZhuuVIP API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LinkUpdate {
  title?: string;
  url?: string;
  icon?: string;
  imageUrl?: string;
  sortOrder?: number;
  isActive?: boolean;
}