# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZhuuVIP API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  profileName?: string;
  profileBio?: string;
  logoUrl?: string;
  bannerUrl?: string;
  themeColor?: string;
  statusText?: string;
}