# lib/api-zod/src/generated/types/songInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZhuuVIP API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SongInput {
  title: string;
  artist: string;
  url: string;
  coverUrl?: string;
  duration?: number;
  sortOrder?: number;
}